Move getRedirectFromPendingPage to lib/pending-page
