+++ client/my-sites/checkout/checkout-thank-you/pending.tsx
@@ -8,9 +8,9 @@ import EmptyContent from 'calypso/components/empty-content';
 import Main from 'calypso/components/main';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import CalypsoShoppingCartProvider from 'calypso/my-sites/checkout/calypso-shopping-cart-provider';
+import { getRedirectFromPendingPage } from 'calypso/my-sites/checkout/composite-checkout/lib/pending-page';
 import useCartKey from 'calypso/my-sites/checkout/use-cart-key';
 import { errorNotice } from 'calypso/state/notices/actions';
-import { SUCCESS, ERROR, FAILURE, UNKNOWN } from 'calypso/state/order-transactions/constants';
 import getOrderTransaction from 'calypso/state/selectors/get-order-transaction';
 import getOrderTransactionError from 'calypso/state/selectors/get-order-transaction-error';
 import type { OrderTransaction } from 'calypso/state/selectors/get-order-transaction';
@@ -90,21 +90,6 @@ function isValidOrderId( orderId: number | ':orderId' ): orderId is number {
 	return Number.isInteger( orderId );
 }
 
-function interpolateReceiptId( url: string, receiptId: number ): string {
-	if ( url.includes( ':receiptId' ) ) {
-		return url.replaceAll( ':receiptId', `${ receiptId }` );
-	}
-
-	// Only treat `/pending` as a placeholder if it's the end of the URL
-	// pathname, but preserve query strings or hashes.
-	const receiptPlaceholderRegexp = /\/pending([?#]|$)/;
-	if ( receiptPlaceholderRegexp.test( url ) ) {
-		return url.replace( receiptPlaceholderRegexp, `/${ receiptId }$1` );
-	}
-
-	return url;
-}
-
 function performRedirect( url: string, siteSlug: string | undefined ): void {
 	if ( url.startsWith( '/' ) ) {
 		page( url );
@@ -155,109 +140,6 @@ function performRedirect( url: string, siteSlug: string | undefined ): void {
 	page( fallbackUrl );
 }
 
-interface RedirectInstructions {
-	url: string;
-	errorNotice?: string;
-}
-
-interface RedirectForTransactionStatusArgs {
-	error?: Error | null;
-	transaction?: OrderTransaction | null;
-	orderId?: number;
-	receiptId?: number;
-	redirectTo?: string;
-	siteSlug?: string;
-	translate: ReturnType< typeof useTranslate >;
-}
-
-function getRedirectForTransactionStatus( {
-	error,
-	transaction,
-	orderId,
-	receiptId,
-	redirectTo,
-	siteSlug,
-	translate,
-}: RedirectForTransactionStatusArgs ): RedirectInstructions | undefined {
-	const defaultFailUrl = siteSlug ? `/checkout/${ siteSlug }` : '/';
-	const defaultFailErrorNotice = translate(
-		"Sorry, we couldn't process your payment. Please try again later."
-	);
-	const planRoute = siteSlug ? `/plans/my-plan/${ siteSlug }` : '/pricing';
-	const defaultSuccessUrl =
-		siteSlug && receiptId
-			? `/checkout/thank-you/${ siteSlug }/${ receiptId }`
-			: '/checkout/thank-you/no-site';
-
-	// If there is a receipt ID, then the order must already be complete. In
-	// that case, we can redirect immediately.
-	if ( receiptId && redirectTo ) {
-		return {
-			url: interpolateReceiptId( redirectTo, receiptId ),
-		};
-	}
-	if ( receiptId && ! redirectTo ) {
-		return {
-			url: interpolateReceiptId( defaultSuccessUrl, receiptId ),
-		};
-	}
-
-	// If the order ID is missing and there is no receiptId, we don't know
-	// what to do because there's no way to know the status of the
-	// transaction. If this happens it's definitely a bug, but let's keep the
-	// existing behavior and send the user to a generic thank-you page,
-	// assuming that the purchase was successful. This goes to the URL path
-	// `/checkout/thank-you/:site/:receiptId` but without a receiptId.
-	if ( ! orderId ) {
-		return {
-			url: `/checkout/thank-you/${ siteSlug ?? 'no-site' }/unknown-receipt`,
-		};
-	}
-
-	if ( transaction ) {
-		const { processingStatus } = transaction;
-
-		// If the order is complete, we can redirect to the final page.
-		if ( SUCCESS === processingStatus ) {
-			const { receiptId: transactionReceiptId } = transaction;
-
-			return {
-				url: interpolateReceiptId( redirectTo ?? defaultSuccessUrl, transactionReceiptId ),
-			};
-		}
-
-		// If the processing status indicates that there was something wrong, it
-		// could be because the user has cancelled the payment, or because the
-		// payment failed after being authorized. Redirect users back to the
-		// checkout page so they can try again.
-		if ( ERROR === processingStatus || FAILURE === processingStatus ) {
-			return {
-				errorNotice: defaultFailErrorNotice,
-				url: defaultFailUrl,
-			};
-		}
-
-		// The API has responded a status string that we don't expect somehow.
-		// Redirect users back to the plan page so that they won't be stuck here.
-		if ( UNKNOWN === processingStatus ) {
-			return {
-				errorNotice: defaultFailErrorNotice,
-				url: planRoute,
-			};
-		}
-	}
-
-	// A HTTP error occured; we will send the user back to checkout.
-	if ( error ) {
-		return {
-			errorNotice: defaultFailErrorNotice,
-			url: defaultFailUrl,
-		};
-	}
-
-	return undefined;
-}
-
 function useRedirectOnTransactionSuccess( {
 	orderId,
 	receiptId,
@@ -291,7 +173,7 @@ function useRedirectOnTransactionSuccess( {
 		// as empty.
 		reloadCart();
 
-		const redirectInstructions = getRedirectForTransactionStatus( {
+		const redirectInstructions = getRedirectFromPendingPage( {
 			error,
 			transaction,
 			orderId,
